// frontend/src/types/project.ts
export interface Project {
  id: number;
  name: string;
  points: number;
  amount: string;
  price: string;
  time: string;
  favorite: boolean;
  stabilityValue: string;
  spreadBps: number;
  Xdays: number;
  mc: string;
  date: string;
  isSpot: boolean;
  isFutures: boolean;
  isTge: boolean;
}